Stop pushing invalid camera device objects to Lua.
`deviceWatcherDoCallback` so that if the `event` string is "Removed",
a nil pushed to the Lua skin instead of a possibly invalid object
`HSCamera` object.

Why?

In `libcamera.m`, when an `AVCaptureDeviceWasDisconnectedNotification`
occurs, the `deviceWatcherDoCallback` function is invoked with the
event type "Removed". Inside this function,
`[cameraManager cameraForDeviceID:deviceId]` is called to retrieve or
create an `HSCamera` object, which is then pushed to the Lua callback.

The problem is that for a "Removed" event, the `CMIODeviceID` refers
to a device that is no longer accessible.

- If `cameraForDeviceID:` returns a cached `HSCamera` object, this
object now represents a disconnected device.

- If it creates a new `HSCamera` object, its `initWithDeviceID:`
method (and subsequently `getCameraUID=/=getCameraName`) will attempt
to query system properties for this disconnected `CMIODeviceID`, which
fails.

In both scenarios, an `HSCamera` object representing an invalid or
non-functional state is pushed to Lua. When Lua attempts to interact
with this object (or even during its flawed internal initialization),
`LuaSkin` errors like "getUID: Unable to get data size" and "Unable to
get camera name for: (null)" are logged, as the underlying native
calls fail.